move providers helpers to a separate files
